Modifying tests
